requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.8.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Image Processing
pillow>=8.0.0
opencv-python>=4.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration
pyyaml>=6.0

# Data Download
kaggle>=1.5.0
kagglehub>=0.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
jupyterlab-widgets>=3.0.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0